+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.30 (master)
 * Limit the jid, name and group name to 256 character max in the Roster to allow it to be saved in the DB
 * Implement XEP-0272: Multiparty Jingle (Muji)
 * Implement XEP-0482: Call Invites
+* Upgrade to fabiang/sasl 2.0
 
 v0.29.2
 ---------------------------